Define what it means to win on rows or columns
Use Iterator.all on all the rows to check for a row that wins.

Project columns and then use an Iterator.all to check for a column that
wins.

We made the Value type have the Copy trait, as it's not that big (an
annotated u32), but if not we could have made it work by making the
column map return a ref:

            if self
                .rows
                .iter()
                .map(|row| &row[c])
                .all(|x| matches!(x, Value::Matched(_)))

In this case we're fine, as the overhead of the enum is the same as the
64-bit pointer we'd be using if we didn't allow the copy:

   std::mem::size_of::<Value>() = 8
   std::mem::size_of::<&Value>() = 8