Day4 Part 2: Implement game.losing_score
Fairly straightfoward, calculate the score of the last card to win.
Only trickiness is in picking up that transition - we fudged it by
keeping a set of the card indexes, and when a board becomes winning we
add it to the set. When the set size == game.boards.len() we know we
just won the last hand.

Something smarter with detecting the edge from !winning to winning would
be more elegant, but this works.