Day 5 Step 2 Compute diagonals
Extract the common function highway_to_the_danger_zones which takes a
function for expanding an Edge into Points.

Had a mild headache as the inbuilt Range is only an increasing range, so
if you want 3..1 you need to generate 1..3 and reverse it.
Iterator::rev() should be the ticket here, but a reversed iterator types
to a Rev<Iterator> so simple punning is defeated.  To work around we
make some temporary vectors, which we can reverse, and then into_iter()
on those.  There's probably a better way to work inside the type system
here, but I didn't see it.