Day 6, Part 1 - Rotating arrays for Population
This is a little more expensive than it could be, we could instead of
rotating the data, store an index for the start of the circular buffer,
then each update is 2 reads, 1 sum, 2 writes rather than the 8 copies, 1
read, 1 sum, 1 write.

At least it's better than the asbolutely naiive version of tracking each
fish.