Clippy guided refactor
It directed me to the 0 as u32 suggesting instead 0_u32.  It turns out
it didn't the type hint though it had in a previous iteration.

Also we can take out the pointer derefence by making the map specify a
reference &v rather than a plain v.  I assme that this is a form of argument
destructuring.