Day8 Part2 decode_segments
We can do a little set logic to narrow the maybe options down.

The values for one and four are enough to get into the other values, and
we know one is len(2) and four is len(4)

Matching gives us a really succinct way to represent this, as we just
construct the triple of (len(s), len(s | one), len(s | four)) and we can
match the values.