Day11 Part1:  Flashing octopuses celluar automata
To make things interesting we pull in the ndarray crate which can do
wonderful things with multi-dimensional arrays.

The simple flahing algorithm is implemented in State.step() which
advanced the internal state and reports on the number of flashes, if
any.

There's a good refactor/code deletion oppourtinity in the bit that adds
energy to the neighbour cells, but that can come later.