Day 12 Part 2: Allow limited doubling back
For this we extende the logic.

Count the number of times we visited all the nodes.  If we've visited
any small cave twice, back to the usual behavior of not returning to
them.

itertools's .count() is just fine for this, though Hashmap returning
references even when the values would be trivially clonable is a bit
ugly.