Day 14 Part 2: Space efficient polymers
Since we explode in size after very few iterations here's a better way
to get the data we want.

Rather than generating a string, the state we move between steps is the
HashMap<(char,char), usize>, the counts for each pair.

This needs a little extra setup to make the pairs, then our per
character counts are taken from the start of the pair, remembering to
add the last chracter one extra time.