Day 17 Part 0 - go crazy implementing a nom parser
Rather than just pulling apart the puzzle input with a regex or similar
we go a little crazy and use a nom parser.

Also we can't seem to stop ourselves from nouning, so we have a
Point,Target, and Probe structs all ready to do the simulation with,
rather than just slinging vectors of tuples around.