Day 19 Part 1 - Mesh the points
This was a real headache.  Ended up cheating and looking online for an
approach -

Maintain a known set of beacons, seeded from scanner 0.
Compute the rotations of the new scanner readings, and then keep the
differences between all those translated points and the known points.
For each of the differences, check if when used as an offset the
translated points match >= 12 of the known points, on success add the
translated points to the known set and return the difference, which is
the location of the scanner.