Day 20 - Part 0
We implement a chunk of the algorithm for the automata.  It's a simple
enough HashSet of lit pixels, so we don't have a bunch of array jiggling
to do when considering the infinity of the field.

We get the correct answer for the example, but not for our actual input,
because we're not handling the inifinite background correctly.

In our puzzle we can see that algorithm[0] = '#', so on step 2 the
entire infinite image outside of our known area will be lit, so we need
to change our bitstring fetcher to be aware of this.