Day20 Part1: Track what the infinite void looks like
To the image we add more state, the bit value for the background and the
known bounds.  After computing the new points, figure out what the
background byte will be by checking the value for some place
out-of-bounds.  When fetching the bit value check if it's in or out of
known bounds, when out of bounds it's 'background', otherwise it's '1',
or '0' depending on if the value has been set.