Day 22: Part 2 - track lit cubes
Rather than simulating the actual lights, we really need to consider the
lit up ranges as a collection cubes.  For each new command, we check if
it intersects with any of the known lit cubes, and if so we slice the
existing cube up into it's 27 subcubes (cubelets?) and discare the
intersection.

Over time this'll mean we end up with a lot of very small cubes
represented - we could go the extra step to fuse them back together to
avoid the complexity of this exploding, but since we only have 420
commands in our inputs we're only talking about tracking 10,490 cubes in
the worst case.