Day 23: Part 0: Initial thoughts
It seems that what we;'re going to need to do is create all the legal
game states, and then search through them for the cheapest way to get
there.

This is pretty much Dijkstra, like we did on day15, but instead of our
costs being tracked in an array, it'll come from the path that the
ampiphods take to attain that state.

The next steps from here are be to come up with a state representation,
and then implement the state machines the ampiphods follow to determine
legal moves.