Declare the graphs of possible moves
Cell is the node type.  Notation for this is {column}{index} with
columns being 'a'..'d' for the home rooms of the A..D amphipods, and
column 'h' being the hallway.

Move has a start, end, set of blocking cells (cells that must be free for
the move to be legal), and the move count.  We declare these for rooms
a0..d0 and then have a routine that derives what those would be for
a1