Day 23 Part 2 - Teach Game to do multiple rows
For our input generator splice in the 'folded' sequence of pods.
Split the input parsing up a little to allow this.

Move the aoc entry points down to the bottom of the file so they're
easier to find.

This whole thing needs some profiling and maybe some retooling.

The day1 solution is found after 22 seconds, day2 hasn't run through yet
but the state space is way bigger.