Switch to an Array representation for State
Cell and State are a little more tightly bound, but state is now a
24-byte array rather than a HashMap<Cell,Pod>.

Pods are now just characters.

Takes part1 down to 1.94 seconds, and should hopefully bring the memory
usage down - we're not really attacking the memory complexity, just the
size in ram of one State.   It's enough to stop us hitting swap -
a ~5 minute runtime but it's completing.

  AOC 2021
  Day 23 - Part 1 : 15365
          generator: 79.571µs,
          runner: 1.941773087s

  Day 23 - Part 2 : 52055
          generator: 75.461µs,
          runner: 325.169495983s