Day 24: Find the largest and smallest model numbers
To test the space somewhat efficiently we break up the MONAD code into
it's sub-progams based on the observation that each subprogram is a
function from (z,w) -> z with the starting state case of (0,z).  We
iterate this map for each successive digit collecting for the maximal
value that goes through that state.